Move Giga downloader into a separate module.
