+++ app/src/main/AndroidManifest.xml
@@ -104,7 +104,7 @@
             android:label="@string/app_name"
             android:launchMode="singleTask"/>
 
-        <service android:name="us.shandian.giga.service.DownloadManagerService"/>
+        <service android:name=".download.service.DownloadManagerService" />
 
         <activity
             android:name=".util.FilePickerActivityHelper"
