--- app/src/main/java/us/shandian/giga/ui/common/ToolbarActivity.java
@@ -1,23 +1,0 @@
-package us.shandian.giga.ui.common;
-
-import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.Toolbar;
-
-import org.schabi.newpipe.R;
-
-public abstract class ToolbarActivity extends AppCompatActivity {
-    protected Toolbar mToolbar;
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(getLayoutResource());
-
-        mToolbar = this.findViewById(R.id.toolbar);
-
-        setSupportActionBar(mToolbar);
-    }
-
-    protected abstract int getLayoutResource();
-}
