--- app/src/main/java/us/shandian/giga/ui/fragment/AllMissionsFragment.java
@@ -1,12 +1,0 @@
-package us.shandian.giga.ui.fragment;
-
-import us.shandian.giga.get.DownloadManager;
-import us.shandian.giga.service.DownloadManagerService;
-
-public class AllMissionsFragment extends MissionsFragment {
-
-    @Override
-    protected DownloadManager setupDownloadManager(DownloadManagerService.DMBinder binder) {
-        return binder.getDownloadManager();
-    }
-}
