+++ downloader/src/main/AndroidManifest.xml
@@ -1,0 +1,1 @@
+<manifest package="org.schabi.newpipe.downloader" />
