+++ downloader/src/main/res/values/colors.xml
@@ -1,0 +1,17 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <!-- GigaGet theme -->
+    <color name="bluegray">#607D8B</color>
+
+    <!-- GigaGet download Theme colors -->
+    <color name="audio_left_to_load_color">#555555</color>
+    <color name="audio_already_load_color">#000000</color>
+    <color name="video_left_to_load_color">#CD5656</color>
+    <color name="video_already_load_color">#BC211D</color>
+
+    <!-- GigaGet Component colors -->
+    <color name="white">#FFFFFF</color>
+    <color name="light_gray">#EFEFEF</color>
+    <color name="middle_gray">#E0E0E0</color>
+    <color name="gray">#616161</color>
+</resources>
